EvoKernel/requirements.txt
numpy
google-cloud-aiplatform
openai
colorama
python-dotenv
requests
beautifulsoup4
duckduckgo-search
psutil